refactor: 태그 리포지토리 추상화
