+++ src/modules/extensions/essay/tag/infrastructure/itag.repository.ts
@@ -1,0 +1,7 @@
+import { Tag } from '../../../../../entities/tag.entity';
+
+export interface ITagRepository {
+  findTag(name: string): Promise<Tag>;
+
+  saveTag(name: string): Promise<Tag>;
+}
