+++ src/modules/extensions/essay/tag/tag.module.ts
@@ -3,12 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../../../../entities/user.entity';
 import { Essay } from '../../../../entities/essay.entity';
 import { Tag } from '../../../../entities/tag.entity';
-import { TagService } from './tag.service';
-import { TagRepository } from './tag.repository';
+import { TagService } from './core/tag.service';
+import { TagRepository } from './infrastructure/tag.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Essay, Tag])],
-  providers: [TagService, TagRepository],
-  exports: [TagService, TagRepository],
+  providers: [TagService, { provide: 'ITagRepository', useClass: TagRepository }],
+  exports: [TagService, { provide: 'ITagRepository', useClass: TagRepository }],
 })
 export class TagModule {}
